requirements.txt
dj-database-url==0.4.1
Django==1.10.2
gunicorn==19.6.0
psycopg2==2.6.2
whitenoise==3.2

# Salesforce Lightning Design System theme
django-slds==0.1.2
django-slds-crispyforms==0.1.0

# Github library
github3.py==0.9.6
